fix: refactor the alert and check-design-name
