+++ src/utils/checkDesignName.ts
@@ -1,6 +1,13 @@
 import { DESIGN_NAME_REGULAR_EXPRESSION } from "@/constants/courtData";
+import { DESIGN_NAME_MAX_CHAR_LENGTH } from "@/constants/courtData";
 
 const checkDesignName = (name: string, nameList: string[]) => {
+  name.trim();
+
+  if (name.length > DESIGN_NAME_MAX_CHAR_LENGTH) {
+    return `The design name should less than ${DESIGN_NAME_MAX_CHAR_LENGTH} characters.`;
+  }
+
   // when canBeBlank is false, check if name is blank
   if (name.length === 0) {
     return "Please enter a design name.";
